refactor: 쿠키 관련 처리 클래스명 변경 및 패키지 이동
