+++ src/main/java/roomescape/config/CheckAdminInterceptor.java
@@ -3,10 +3,10 @@ package roomescape.config;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.web.servlet.HandlerInterceptor;
+import roomescape.auth.CookieProvider;
 import roomescape.domain.Member;
 import roomescape.domain.Role;
 import roomescape.service.AuthService;
-import roomescape.utils.CookieUtils;
 
 public class CheckAdminInterceptor implements HandlerInterceptor {
     private final AuthService authService;
@@ -17,7 +17,7 @@ public class CheckAdminInterceptor implements HandlerInterceptor {
 
     @Override
     public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) {
-        final String token = CookieUtils.extractTokenFrom(request);
+        final String token = CookieProvider.extractTokenFrom(request);
         final Member member = authService.findMemberByToken(token);
 
         if (member == null || isUser(member.getRole())) {
