+++ src/main/java/roomescape/config/LoginMemberArgumentResolver.java
@@ -7,8 +7,8 @@ import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
+import roomescape.auth.CookieProvider;
 import roomescape.service.AuthService;
-import roomescape.utils.CookieUtils;
 
 @Component
 public class LoginMemberArgumentResolver implements HandlerMethodArgumentResolver {
@@ -27,7 +27,7 @@ public class LoginMemberArgumentResolver implements HandlerMethodArgumentResolve
     public Object resolveArgument(MethodParameter parameter, ModelAndViewContainer mavContainer,
                                   NativeWebRequest webRequest, WebDataBinderFactory binderFactory) {
         final HttpServletRequest request = (HttpServletRequest) webRequest.getNativeRequest();
-        final String token = CookieUtils.extractTokenFrom(request);
+        final String token = CookieProvider.extractTokenFrom(request);
 
         return authService.findMemberByToken(token);
     }
